+++ CHANGES.markdown
@@ -3,6 +3,7 @@
  - `Functor`. `Applicative`, `Alternative`, and `MonadPlus` instances for
    `ArrowMonad`
  - Expose `Read` and `Show` instances for `Down` on GHCs before 7.8
+ - `Bits`, `Bounded`, and `Integral` instances for `CDev`
 
 ## Changes in 0.4.1
  - Fixed imports on GHC < 7.8 on Windows